(update) extract system nodeDefs from nodeDefStore
