+++ src/stores/data.ts
@@ -1,0 +1,37 @@
+import { ComfyNodeDef } from '@/types/apiTypes'
+
+export const SYSTEM_NODE_DEFS: ComfyNodeDef[] = [
+  {
+    name: 'PrimitiveNode',
+    display_name: 'Primitive',
+    category: 'utils',
+    input: { required: {}, optional: {} },
+    output: ['*'],
+    output_name: ['connect to widget input'],
+    output_is_list: [false],
+    python_module: 'nodes',
+    description: 'Primitive values like numbers, strings, and booleans.'
+  },
+  {
+    name: 'Reroute',
+    display_name: 'Reroute',
+    category: 'utils',
+    input: { required: { '': ['*'] }, optional: {} },
+    output: ['*'],
+    output_name: [''],
+    output_is_list: [false],
+    python_module: 'nodes',
+    description: 'Reroute the connection to another node.'
+  },
+  {
+    name: 'Note',
+    display_name: 'Note',
+    category: 'utils',
+    input: { required: {}, optional: {} },
+    output: [],
+    output_name: [],
+    output_is_list: [],
+    python_module: 'nodes',
+    description: 'Node that add notes to your project'
+  }
+]
